requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
configobj==5.0.6
configparser==5.3.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
etelemetry==0.3.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fitz==0.0.1.dev2
Flask==2.2.2
future==0.18.2
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
looseversion==1.0.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nibabel==4.0.2
nipype==1.8.5
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdf2image==1.16.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.32
prov==2.0.0
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
PyMuPDF==1.21.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pywinpty==2.0.9
pyxnat==1.5
pyzmq==24.0.1
rdflib==6.2.0
requests==2.28.1
scipy==1.9.3
Send2Trash==1.8.0
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
terminado==0.17.0
tesseract==0.1.3
tinycss2==1.2.1
tornado==6.2
traitlets==5.5.0
traits==6.3.2
urllib3==1.26.12
virtualenv==20.16.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
